pair worked with emma to connect the sellers username a ref to the sellers profile

+++ src/main/java/com/codeup/adlister/dao/Users.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface Users {
     User findByUsername(String username);
-    User findByUserId(int user_id);
+    User findByUserId(long user_id);
     Long insert(User user);
     boolean delete(long id);
     boolean update(User user);
